- performance of fail2ban optimized    -- cache dnsToIp, ipToName to prevent long wait during retrieving of ip/name for wrong dns or lazy dns-system;    -- instead of simple "sleep" used conditional wait "wait_for", that internal increases sleep interval up to sleeptime;    -- ticket / banmanager / failmanager modules are performance optimized; - performance of test cases optimized:    -- added option "--fast" to decrease wait intervals, avoid passive waiting, and skip few very slow test cases; - code review after partially cherry pick of branch 'ban-time-incr' (see gh-716)    -- ticket module prepared to easy merge with newest version of 'ban-time-incr', now additionally holds banTime, banCount and json-data;    -- executeCmd partially moved from action to new module utils, etc.    -- python 2.6 compatibility;

+++ fail2ban/server/transmitter.py
@@ -95,7 +95,7 @@ class Transmitter:
 			return None
 		elif command[0] == "sleep":
 			value = command[1]
-			time.sleep(int(value))
+			time.sleep(float(value))
 			return None
 		elif command[0] == "flushlogs":
 			return self.__server.flushLogs()
